fix (3_Factorial_Calculation): fix cpp code to return big number

+++ Beginner/3_Factorial_Calculation/my_solution/cpp/factorial_calculation.hpp
@@ -3,6 +3,6 @@
 
 // Prototype definition
 
-int factorialCalculation(int index);
+long long int factorialCalculation(int index);
 
 #endif
